- **Feature**: Comprehensive `FieldProperties` expansion with 20+ new properties

+++ app/src/main/java/com/neome/api/meta/base/dto/FieldValueDecimal.kt
@@ -5,7 +5,6 @@
 
 package com.neome.api.meta.base.dto
 
-interface FieldValueDecimal
-{
-  val value: Long?
+interface FieldValueDecimal {
+    val value: Double?
 }
